feat(diagnoses): adds ability to view a diagnosis
In the Diagnoses tab, the diagnosis list was replaced with a table containing
the name and dates for the diagnoses, along with a button that allows a
read-permission user to view each diagnosis form in its entirety. The same
routing is also applied to the smaller diagnosis table above all of the patient
tabs. Some of the corresponding test files were added or changed to accommodate
for this new feature.